requirements.txt
alembic==1.7.6
aniso8601==9.0.1
attrs==21.4.0
click==8.0.3
DateTime==4.3
Flask==2.0.2
Flask-HTTPAuth==4.5.0
Flask-Login==0.5.0
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.10.1
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.4.0
Mako==1.1.6
MarkupSafe==2.0.1
passlib==1.7.4
psycopg2-binary==2.9.3
pyrsistent==0.18.1
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.31
Werkzeug==2.0.3
zipp==3.7.0
zope.interface==5.4.0
